pyproject.toml
[tool.poetry]
name = "ts-channel_django-temp"
version = "0.1.0"
description = ""
authors = ["matheus <matheus.tessaroli@olist.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
psycopg2-binary = "^2.8.6"
pytest-html = "^3.1.1"
coverage = "^5.5"
pre-commit = "^2.13.0"
pylint = "^2.8.3"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"
pytest = "^6.2.4"
pytest-django = "^4.4.0"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
